norme pour textures, move et parse_texture
