+++ src/textures.c
@@ -12,6 +12,27 @@
 
 #include "cub3d.h"
 
+bool	load_texture(t_game *game, t_texture *texture, char *path)
+{
+	while (*path && (*path == ' ' || *path == '\t'))
+		path++;
+	texture->img = mlx_xpm_file_to_image(game->mlx, path, &texture->width, &texture->height);
+	if (!texture->img)
+	{
+		printf("Error\nFailed to load texture: %s\n", path);
+		return (false);
+	}
+	texture->addr = mlx_get_data_addr(texture->img, &texture->bits_per_pixel, &texture->line_length, &texture->endian);
+	if (texture->width != 64 || texture->height != 64)
+	{
+		printf("Error\nThe texture %s is not 64x64 pixels (current dimensions: %dx%d)\n", path, texture->width, texture->height);
+		mlx_destroy_image(game->mlx, texture->img);
+		texture->img = NULL;
+		return (false);
+	}
+	return (true);
+}
+
 int	get_tex_pixel(t_texture *texture, int x, int y)
 {
 	char	*dst;
